==> part1_calculations/step1_preprocessing.ipynb <==
import pickle
import pandas as pd
import datetime
import numpy as np

### Getting all the focal brands

top_brands = pd.read_csv('../data/top_brands_by_category.csv')
top_brands.head()

focal_brands = top_brands['BRANDS'].to_list()
focal_brands

### Getting the brands catalog

brands_catalog = pd.read_csv('../data/catalog.tsv', sep='\t', header=None, names=['ID', 'Name', 'Genre', 'Type', 'Classification', 'Status'])
brands_catalog['Name_Standard'] = brands_catalog['Name'].apply(lambda x: x.strip().lower()) # For comparison with visitation_data.csv
brands_catalog.head()

### Getting the brand visitation data

brands_visits = pd.read_csv('../data/revision_visits_revenue_2019.csv')
brands_visits['brand_standard'] = brands_visits['brand'].apply(lambda x: x.strip().lower()) # For comparison with catalog.tsv
brands_visits['date'] = brands_visits['date'].apply(lambda x: datetime.datetime.strptime(x, '%Y-%m-%d').date())
brands_visits = brands_visits.rename(columns={'brand': 'brand_visitation'})
brands_visits.head()

print('Unique Brands having visitation data: ', len(brands_visits['brand_standard'].unique()))

Calculating all the brands which have visitation data, and they are present in the catalog as well (have a valid ID which can be used to fetch the social data)

brand_catalog_visits = pd.merge(left=brands_visits, right=brands_catalog, how="left", left_on="brand_standard", right_on="Name_Standard")
brand_catalog_visits.head()

brand_catalog_visits = brand_catalog_visits[~pd.isna(brand_catalog_visits['ID'])]

print('Unique Brands having visitation data and Valid ID from catalog: ', len(brand_catalog_visits['Name_Standard'].unique()))

### Getting the social data

social_data = pd.read_csv('../data/social.csv')
social_data['date'] = social_data['date'].apply(lambda x: datetime.datetime.strptime(x, '%Y-%m-%d').date())
social_data.head()

Getting all the unique IDs of brands who have visitation data and a respective ID in brands catalog

brand_catalog_visits_unique_id = list(np.int64(brand_catalog_visits['ID'].unique()))

Getting the social data for brands which have visitation data and catalog ID

social_data = social_data[social_data['id'].isin(brand_catalog_visits_unique_id)]

print('Unique brands having social data, visitation data and Valid ID from catalog: ', len(social_data['id'].unique()))

len(social_data)

Saving all the important dataframes as pickle files to be loaded again

with open('top_brands.pickle', 'wb') as file:
    pickle.dump(top_brands, file)

with open('brand_catalog_visits.pickle', 'wb') as file:
    pickle.dump(brand_catalog_visits, file)

with open('social_data.pickle', 'wb') as file:
    pickle.dump(social_data, file)
