==> part1_calculations/step2_calculations.ipynb <==
This program will read the filtered data, and peform the calculations for one focal brand

import pickle
import datetime
import pandas as pd
import numpy as np
from tqdm import tqdm
pd.options.mode.chained_assignment = None  # default='warn'

with open('top_brands.pickle', 'rb') as file:
    top_brands = pickle.load(file)

top_brands.head()

with open('brand_catalog_visits.pickle', 'rb') as file:
    brand_catalog_visits = pickle.load(file)

brand_catalog_visits.head()

with open('social_data.pickle', 'rb') as file:
    social_data = pickle.load(file)

social_data.head()

### Calculating the localized number of reviews for all brands having social data available

Function to perform calculations using Step 2

def get_brand_visit_data(brand_id):
    brand_visit_data = None
    brand_social_data = None
    
    brand_visit_data = brand_catalog_visits[brand_catalog_visits['ID'] == brand_id]
    brand_social_data = social_data[social_data['id'] == brand_id]
    
    brand_visit_data.loc[:,'visits_past_60_days'] = brand_visit_data.groupby(by=['lat', 'lon'], group_keys=False)[['lat', 'lon', 'date', 'visits_by_day']].apply(lambda x: x.sort_values('date').rolling(60)['visits_by_day'].sum().shift(1)).squeeze()
    brand_visit_data.loc[:,'visits_past_3_days'] = brand_visit_data.groupby(by=['lat', 'lon'], group_keys=False)[['lat', 'lon', 'date', 'visits_by_day']].apply(lambda x: x.sort_values('date').rolling(3)['visits_by_day'].sum().shift(1)).squeeze()
    
    brand_visit_data.loc[:,'total_visits_across_stores_60_days'] = brand_visit_data.groupby(by=['date'])['visits_past_60_days'].transform('sum')
    brand_visit_data.loc[:,'proportion_of_visits_60_days'] = brand_visit_data['visits_past_60_days']/brand_visit_data['total_visits_across_stores_60_days']
    
    def get_localized_fb(df_rec):
        localized_fb_reviews = None
        soc_data = brand_social_data[brand_social_data['date'] == df_rec.name]
        
        if (len(soc_data) != 0):    
            localized_fb_reviews = soc_data['num_review_fb'].values[0] * df_rec
        else:
            localized_fb_reviews = np.nan * df_rec
        
        return localized_fb_reviews
    
    def get_localized_ig(df_rec):
        localized_ig_reviews = None
        soc_data = brand_social_data[brand_social_data['date'] == df_rec.name]
        
        if (len(soc_data) != 0):    
            localized_ig_reviews = soc_data['num_review_ig'].values[0] * df_rec
        else:
            localized_ig_reviews = np.nan * df_rec
        
        return localized_ig_reviews
    
    def get_localized_tw(df_rec):
        localized_tw_reviews = None
        soc_data = brand_social_data[brand_social_data['date'] == df_rec.name]
        
        if (len(soc_data) != 0):    
            localized_tw_reviews = soc_data['num_review_tw'].values[0] * df_rec
        else:
            localized_tw_reviews = np.nan * df_rec
        
        return localized_tw_reviews
    
    brand_visit_data.loc[:,'localized_fb_reviews_60_days'] = brand_visit_data.groupby(by=['date'])['proportion_of_visits_60_days'].apply(get_localized_fb).droplevel(0)
    brand_visit_data.loc[:,'localized_ig_reviews_60_days'] = brand_visit_data.groupby(by=['date'])['proportion_of_visits_60_days'].apply(get_localized_ig).droplevel(0)
    brand_visit_data.loc[:,'localized_tw_reviews_60_days'] = brand_visit_data.groupby(by=['date'])['proportion_of_visits_60_days'].apply(get_localized_tw).droplevel(0)
    
    return brand_visit_data

valid_brand_ids = social_data['id'].unique().tolist()

brand_visit_local_reviews = None

for i in tqdm(range(0, len(valid_brand_ids))):    
    brand_data_df = get_brand_visit_data(valid_brand_ids[i])

    if i == 0: # For the very first time, store the dataframe. For any next iterations, just perform the concatenation
        brand_visit_local_reviews = brand_data_df
    else:
        brand_visit_local_reviews = pd.concat([brand_visit_local_reviews, brand_data_df], axis=0)

with open('brand_visit_local_reviews.pickle', 'wb') as file:
    pickle.dump(brand_visit_local_reviews, file)
